+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 * refactor the SQLAlchemy Bundle to split most of it out into its own package, so that it can be used on its own (without Flask).
 * fix the resource url prefix descriptor to convert to kebab-case instead of snake-case
 * rename `Controller.template_folder` to `Controller.template_folder_name`
+* add `Controller.make_response` as an alias for `flask.make_response`
 
 ## 0.5.1 (2018/07/25)